Move ResultImpl to the util package
