+++ core/src/main/java/com/scalar/db/storage/cosmos/ResultInterpreter.java
@@ -11,7 +11,7 @@ import com.scalar.db.io.FloatValue;
 import com.scalar.db.io.IntValue;
 import com.scalar.db.io.TextValue;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.common.ResultImpl;
+import com.scalar.db.util.ResultImpl;
 import java.util.Base64;
 import java.util.HashMap;
 import java.util.List;
